# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /imports/atlassian/configs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get import config (admin only)
      operationId: getExternalImportConfig
      responses:
        '200':
          description: Config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      summary: Update import config (admin only)
      operationId: updateExternalImportConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExternalImportConfigDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete import config (admin only)
      operationId: deleteExternalImportConfig
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /tickets:
    get:
      summary: List project tickets
      description: |
        Returns tickets for workspaces the user can access; global admins see all.
        For realtime list coherence after mutations, subscribe to the Socket.IO **tickets** namespace
        (see AsyncAPI `tickets/*` channels in `spec/asyncapi.yaml`); REST remains the source of truth for loads.
      operationId: listTickets
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, todo, in_progress, prototype, done, closed]
        - in: query
          name: parentId
          schema:
            type: string
          description: Use literal `null` for root tickets only
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across title, content, status, priority, id
      responses:
        '200':
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponseDto'
        '403':
          description: Forbidden
    post:
      summary: Create ticket
      description: |
        Optional `creationTemplate` is request-only and not stored. Value `specification` creates four spec-driven subtasks
        under the new root ticket in one transaction; invalid with `parentId` (400).
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTicketResponseDto'
  /tickets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get ticket
      operationId: getTicket
      parameters:
        - in: query
          name: includeDescendants
          schema:
            type: boolean
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponseDto'
        '404':
          description: Not found
    patch:
      summary: Update ticket
      operationId: updateTicket
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketDto'
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponseDto'
    delete:
      summary: Delete ticket
      operationId: deleteTicket
      parameters:
        - in: query
          name: releaseExternalSyncMarker
          required: false
          schema:
            type: boolean
            default: false
          description: >
            When true, deletes external import sync markers tied to this ticket so the same remote item may be imported again.
            When false (default), markers are kept with a null local id so scheduled import skips re-creating the row until markers are cleared.
      responses:
        '204':
          description: Deleted
        '409':
          description: Has subtasks
  /tickets/{id}/prototype-prompt:
    get:
      summary: Build prototype prompt for ticket tree
      operationId: getTicketPrototypePrompt
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Prototype prompt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrototypePromptResponseDto'
  /tickets/{id}/comments:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List ticket comments
      operationId: listTicketComments
      responses:
        '200':
          description: Ticket comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketCommentResponseDto'
    post:
      summary: Add ticket comment
      operationId: addTicketComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketCommentDto'
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketCommentResponseDto'
  /tickets/{id}/activity:
    get:
      summary: List ticket activity
      operationId: listTicketActivity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Ticket activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketActivityResponseDto'
  /tickets/{id}/body-generation-sessions:
    post:
      summary: Start AI body generation session
      operationId: startTicketBodyGenerationSession
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartBodyGenerationSessionDto'
      responses:
        '200':
          description: AI body generation session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartBodyGenerationSessionResponseDto'
  /tickets/{id}/apply-generated-body:
    post:
      summary: Apply AI-generated ticket body
      operationId: applyTicketGeneratedBody
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyGeneratedBodyDto'
      responses:
        '200':
          description: AI-generated ticket body applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponseDto'
  /tickets/{id}/migrate:
    post:
      summary: Migrate ticket subtree to another workspace
      description: |
        Moves the resolved root ticket and all of its descendant tickets that share the source workspace to
        `targetClientId`. The caller may pass any ticket id in the subtree; the server always migrates the full tree.
        Requires workspace management (see `canManageWorkspaceConfiguration` on clients) on **both** the source and
        target workspaces. Emits `ticketRemoved` for each id on the source workspace room and `ticketUpsert` on the
        target workspace room (tickets Socket.IO namespace). `targetClientId` must differ from the ticket workspace (400 otherwise).
      operationId: migrateTicketToWorkspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrateTicketDto'
      responses:
        '200':
          description: Migrated root ticket (includes `children` when descendants exist)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrateTicketResponseDto'
        '400':
          description: Invalid request (e.g. target workspace equals source)
        '403':
          description: Missing workspace access or insufficient workspace role for source or target
        '404':
          description: Ticket not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    TicketPriority:
      type: string
      enum: [low, medium, high, critical]
    TicketStatus:
      type: string
      enum: [draft, todo, in_progress, prototype, done, closed]
    TicketActorType:
      type: string
      enum: [human, ai, system]
    TicketActionType:
      type: string
      description: Ticket activity row discriminator; stored as varchar(64) in the database.
      enum:
        - CREATED
        - FIELD_UPDATED
        - STATUS_CHANGED
        - PRIORITY_CHANGED
        - WORKSPACE_MOVED
        - PARENT_CHANGED
        - DELETED
        - COMMENT_ADDED
        - CONTENT_APPLIED_FROM_AI
        - BODY_GENERATION_STARTED
        - PROTOTYPE_PROMPT_GENERATED
        - AUTOMATION_CLAIMED
        - AUTOMATION_STARTED
        - AUTOMATION_SUCCEEDED
        - AUTOMATION_FAILED
        - AUTOMATION_TIMED_OUT
        - AUTOMATION_ESCALATED
        - AUTOMATION_REQUEUED
        - AUTOMATION_APPROVAL_INVALIDATED
        - AUTOMATION_CANCELLED
        - AUTOMATION_APPROVED
        - AUTOMATION_UNAPPROVED
        - AUTOMATION_ELIGIBILITY_CHANGED
        - AUTOMATION_APPROVAL_REQUIREMENT_CHANGED
        - AUTOMATION_SETTINGS_UPDATED
    TicketMarkdownCheckboxTaskCountsDto:
      type: object
      required: [open, done]
      properties:
        open:
          type: integer
          minimum: 0
          description: Count of `[ ]` style open task checkboxes in the described scope.
        done:
          type: integer
          minimum: 0
          description: Count of `[x]` or `[X]` done checkboxes in the described scope.
    TicketTasksDto:
      type: object
      required: [open, done, children]
      properties:
        open:
          type: integer
          minimum: 0
          description: Open checkboxes in this ticket's `content`.
        done:
          type: integer
          minimum: 0
          description: Done checkboxes in this ticket's `content`.
        children:
          description: |
            Sums of open/done checkboxes on all descendant tickets (direct and nested), excluding this ticket's body.
          $ref: '#/components/schemas/TicketMarkdownCheckboxTaskCountsDto'
    TicketShasDto:
      type: object
      required: [short, long]
      properties:
        short:
          type: string
          description: Stable short ticket reference derived from ticket id hash.
        long:
          type: string
          description: Stable full ticket reference derived from ticket id hash.
    TicketResponseDto:
      type: object
      required:
        [
          id,
          shas,
          clientId,
          title,
          priority,
          status,
          automationEligible,
          createdAt,
          updatedAt,
          tasks,
        ]
      properties:
        id:
          type: string
          format: uuid
        shas:
          $ref: '#/components/schemas/TicketShasDto'
        clientId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        createdByUserId:
          type: string
          format: uuid
        createdByEmail:
          type: string
        preferredChatAgentId:
          type: string
          format: uuid
          description: Preferred workspace agent for chat/AI when viewing this ticket (agent-manager agent id).
        automationEligible:
          type: boolean
          description: Whether autonomous prototyping is enabled for this ticket (ticket_automation.eligible).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tasks:
          $ref: '#/components/schemas/TicketTasksDto'
        children:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponseDto'
    CreateTicketDto:
      type: object
      required: [title]
      properties:
        clientId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        creationTemplate:
          type: string
          enum: [empty, specification]
          description: |
            Not persisted. `specification` only allowed when creating a root ticket (omit `parentId`);
            creates Proposal, Specifications, Technical design, and Implementation plan subtasks.
    CreateTicketResponseDto:
      allOf:
        - $ref: '#/components/schemas/TicketResponseDto'
        - type: object
          properties:
            createdChildTickets:
              type: array
              description: Present when `creationTemplate` was `specification` (four subtasks).
              items:
                $ref: '#/components/schemas/TicketResponseDto'
    MigrateTicketDto:
      type: object
      required: [targetClientId]
      properties:
        targetClientId:
          type: string
          format: uuid
          description: Destination workspace (client) id
    MigrateTicketResponseDto:
      type: object
      required: [ticket]
      properties:
        ticket:
          $ref: '#/components/schemas/TicketResponseDto'
          description: Root ticket after migration, including nested `children` when present
    UpdateTicketDto:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        preferredChatAgentId:
          type: string
          format: uuid
          description: Persist preferred workspace agent for chat/AI; send null to clear.
    TicketCommentResponseDto:
      type: object
      required: [id, ticketId, body, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ticketId:
          type: string
          format: uuid
        authorUserId:
          type: string
          format: uuid
        authorEmail:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateTicketCommentDto:
      type: object
      required: [body]
      properties:
        body:
          type: string
    TicketActivityResponseDto:
      type: object
      required: [id, ticketId, occurredAt, actorType, actionType, payload]
      properties:
        id:
          type: string
          format: uuid
        ticketId:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        actorType:
          $ref: '#/components/schemas/TicketActorType'
        actorUserId:
          type: string
          format: uuid
        actorEmail:
          type: string
        actionType:
          $ref: '#/components/schemas/TicketActionType'
        payload:
          type: object
          additionalProperties: true
          description: >
            Shape depends on actionType (e.g. automation eligibility changes include { eligible: boolean };
            AUTOMATION_SETTINGS_UPDATED includes { fields: string[] };
            AUTOMATION_APPROVAL_INVALIDATED may include { reason: string, fields: string[] }).
...
    PrototypePromptResponseDto:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
    StartBodyGenerationSessionDto:
      type: object
      properties:
        agentId:
          type: string
          format: uuid
    StartBodyGenerationSessionResponseDto:
      type: object
      required: [generationId, expiresAt, activity]
      properties:
        generationId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        activity:
          $ref: '#/components/schemas/TicketActivityResponseDto'
    ApplyGeneratedBodyDto:
      type: object
      required: [generationId, content]
      properties:
        generationId:
          type: string
          format: uuid
        content:
          type: string
...
    ExternalImportProviderId:
      type: string
      enum: [atlassian]
    ExternalImportKind:
      type: string
      enum: [jira, confluence]
    ExternalImportConfigResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          $ref: '#/components/schemas/ExternalImportProviderId'
        importKind:
          $ref: '#/components/schemas/ExternalImportKind'
        atlassianConnectionId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        enabled:
          type: boolean
        jiraBoardId:
          type: integer
        jql:
          type: string
        importTargetTicketStatus:
          $ref: '#/components/schemas/TicketStatus'
        confluenceSpaceKey:
          type: string
        confluenceRootPageId:
          type: string
        cql:
          type: string
        agenstraParentTicketId:
          type: string
          format: uuid
        agenstraParentFolderId:
          type: string
          format: uuid
        lastRunAt:
          type: string
          format: date-time
        lastError:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateExternalImportConfigDto:
      type: object
      properties:
        atlassianConnectionId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        enabled:
          type: boolean
        jiraBoardId:
          type: integer
        jql:
          type: string
        importTargetTicketStatus:
          $ref: '#/components/schemas/TicketStatus'
        confluenceSpaceKey:
          type: string
        confluenceRootPageId:
          type: string
        cql:
          type: string
        agenstraParentTicketId:
          type: string
          format: uuid
        agenstraParentFolderId:
          type: string
          format: uuid
...